WebHost: add a page to manage session cookie
