+++ WebHostLib/templates/siteMap.html
@@ -26,6 +26,7 @@
             <li><a href="/user-content">User Content</a></li>
             <li><a href="{{url_for('stats')}}">Game Statistics</a></li>
             <li><a href="/glossary/en">Glossary</a></li>
+            <li><a href="{{url_for("show_session")}}">Session / Login</a></li>
         </ul>
 
         <h2>Tutorials</h2>
